Release v0.3.5: float fps and a release tag guard

+++ src/mover/pipeline.py
@@ -240,7 +240,7 @@ class AnimationPipeline(BasePipeline):
                 int(server_config['port']),
                 server_config['create_video'],
                 output_format=server_config.get('output_format', 'mp4'),
-                video_fps=int(server_config.get('video_fps', DEFAULT_FPS)),
+                video_fps=float(server_config.get('video_fps', DEFAULT_FPS)),
                 capture_duration=server_config.get('capture_duration'),
             )
